Remove unnecessary page components & move components that call state to containers folder
