+++ README.md
@@ -28,7 +28,7 @@
 
 ### `npm start`
 
-Runs the app in the development mode.\
+Runs the Cook This! app in the development mode.\
 
 ### `npm test`
 
@@ -37,8 +37,7 @@ Launches the test runner in the interactive watch mode.\
 
 ### `npm run build`
 
-Builds the app for production to the `build` folder.\
-It correctly bundles React in production mode and optimizes the build for the best performance.
+Builds the Cook This! app for production to the `public` folder.\
 
 
 ### Build With
