--- src/pages/Main.js
@@ -1,9 +1,0 @@
-import RecipeList from '../components/recipeList/RecipeList';
-
-const Main = () => (
-  <div>
-    <RecipeList />
-  </div>
-);
-
-export default Main;
