--- src/pages/Recipe.js
@@ -1,9 +1,0 @@
-import RecipeItem from '../components/recipe/RecipeItem';
-
-const Recipe = () => (
-  <div>
-    <RecipeItem />
-  </div>
-);
-
-export default Recipe;
